+++ t/634-bite-email-postfix.t
@@ -41,6 +41,14 @@ my $isexpected = [
     { 'n' => '43', 's' => qr/\A4[.]3[.]0\z/,    'r' => qr/mailererror/,'b' => qr/\A1\z/ },
     { 'n' => '44', 's' => qr/\A5[.]7[.]1\z/,    'r' => qr/norelaying/, 'b' => qr/\A1\z/ },
     { 'n' => '45', 's' => qr/\A4[.]3[.]0\z/,    'r' => qr/mailboxfull/,'b' => qr/\A1\z/ },
+    { 'n' => '46', 's' => qr/\A5[.]0[.]0\z/,    'r' => qr/userunknown/,'b' => qr/\A0\z/ },
+    { 'n' => '47', 's' => qr/\A5[.]0[.]0\z/,    'r' => qr/systemerror/,'b' => qr/\A1\z/ },
+    { 'n' => '48', 's' => qr/\A5[.]0[.]0\z/,    'r' => qr/toomanyconn/,'b' => qr/\A1\z/ },
+    { 'n' => '49', 's' => qr/\A4[.]0[.]0\z/,    'r' => qr/blocked/,    'b' => qr/\A1\z/ },
+    { 'n' => '50', 's' => qr/\A4[.]0[.]0\z/,    'r' => qr/blocked/,    'b' => qr/\A1\z/ },
+    { 'n' => '51', 's' => qr/\A5[.]7[.]0\z/,    'r' => qr/policyviolation/,'b' => qr/\A1\z/ },
+    { 'n' => '52', 's' => qr/\A5[.]0[.]0\z/,    'r' => qr/suspend/,    'b' => qr/\A1\z/ },
+    { 'n' => '53', 's' => qr/\A5[.]0[.]0\z/,    'r' => qr/syntaxerror/,'b' => qr/\A1\z/ },
 ];
 
 $enginetest->($enginename, $isexpected);