+++ t/643-bite-email-x2.t
@@ -11,6 +11,7 @@ my $isexpected = [
     { 'n' => '02', 's' => qr/\A5[.]0[.]\d+\z/, 'r' => qr/(?:filtered|suspend)/, 'b' => qr/\A1\z/ },
     { 'n' => '03', 's' => qr/\A5[.]0[.]\d+\z/, 'r' => qr/expired/,     'b' => qr/\A1\z/ },
     { 'n' => '04', 's' => qr/\A5[.]0[.]\d+\z/, 'r' => qr/mailboxfull/, 'b' => qr/\A1\z/ },
+    { 'n' => '05', 's' => qr/\A4[.]1[.]9\z/,   'r' => qr/expired/,     'b' => qr/\A1\z/ },
 ];
 
 $enginetest->($enginename, $isexpected);