+++ xt/634-bite-email-postfix.t
@@ -204,6 +204,14 @@ my $isexpected = [
     { 'n' => '01194', 'r' => qr/userunknown/    },
     { 'n' => '01195', 'r' => qr/expired/        },
     { 'n' => '01196', 'r' => qr/userunknown/    },
+    { 'n' => '01197', 'r' => qr/userunknown/    },
+    { 'n' => '01198', 'r' => qr/systemerror/    },
+    { 'n' => '01199', 'r' => qr/toomanyconn/    },
+    { 'n' => '01200', 'r' => qr/blocked/        },
+    { 'n' => '01201', 'r' => qr/blocked/        },
+    { 'n' => '01202', 'r' => qr/policyviolation/ },
+    { 'n' => '01203', 'r' => qr/suspend/        },
+    { 'n' => '01204', 'r' => qr/syntaxerror/    },
 ];
 
 plan 'skip_all', sprintf("%s not found", $samplepath) unless -d $samplepath;